Upgrade to Spring Boot 4.0.0-RC1
Testing milestones is now easier as they are released on Maven Central.

With the upgrade to Spring Boot 4, the monolithic auto-configuration
modules are now split in modules per technology so that features are
opt-in rather than relying on the presence of necessary classes on the
classpath.

The web starter no longer provides an HTTP client so our use of the
RestTemplate fails the build. To opt-in for a rest client, we can
add the 'spring-boot-starter-restclient' starter. Modularization means
that new modules follow the 'spring-boot-<technology>' pattern and
starters are also added for consistency.

The same approach is applied to tests. In fact, each starter has a
`-test` companion for consistency that brings the test infrastructure.
As a result, 'spring-boot-starter-test' is not really used directly
just like 'spring-boot-starter' wasn't in previous Spring Boot
generations.

New modules also mean that the package structure has been harmonized to
match them. We need to fix the import to a few classes.

Our project now builds. For completeness, this commit also renames the
web starter to "webmvc". Now that starters are more "narrowed", we took
the opportunity to clarify their scope. 'webmvc' is now at the same
level as 'webflux'.