Enable API versioning
We want to introduce a new endpoint that provides statistics about the
queue of players, and keep a record of what the API was initially. API
versioning allows us to tag our endpoint easily and specify how
consumers provides the version they want to use.

This commit also tests the new endpoint with RestTestClient, recently
introduced as an alternative to WebTestClient for imperative scenarios.