Enable resilient methods
Now that we have identified that the stats endpoint is causing trouble,
we can make the call to the backend resilient by adding @Retryable.

Spring Framework 7 adds core resiliency features, inspired by the Spring
Retry project that is now superseded by it. @Retryable lets you
configure which exceptions should lead to a retry, and the retry policy
to use.